Se modifico el tipo de dato del Monto en los modelos de los reportes. Ademas se modificaron los casteos en el catalogo de pedidos de los campos anteriores.

+++ Modelos/ModeloReporteEncabezado.cs
@@ -15,6 +15,6 @@ namespace Modelos
        public string PersonaHasta { get; set; }
        public List<ModeloReportePedidoEntreFechas> detallePedido { get; set; }
        public List<ModeloReporteVentaEntreFechas> detalleVenta { get; set; }
-       public float MontoTotal { get; set; }
+       public decimal MontoTotal { get; set; }
     }
 }
